Fix building of tests

+++ UnitTests/CKUploaderTests.m
@@ -93,7 +93,7 @@
 
 - (void)uploader:(CKUploader *)uploader appendString:(NSString *)string toTranscript:(CK2TranscriptType)transcript
 {
-    NSLog(@"%d: %@", transcript, string);
+    NSLog(@"%d: %@", (int)transcript, string);
 }
 
 #pragma mark - Utilities
